requirements.txt
accelerate==1.7.0
diffusers==0.34.0
hf-transfer
hf-xet
huggingface-hub
numpy>=1.24.0
pillow
runpod==1.7.9
safetensors==0.5.3
scipy
setuptools
torch==2.7.0
torchvision
transformers==4.52.4
triton
xformers==0.0.30
# Video generation dependencies for Wan2.1
ftfy
opencv-python
imageio
# Firebase cloud storage and database
firebase-admin>=6.0.0
imageio-ffmpeg